Update Delete order page and button
Update secondary button to have a type or Primary, Secondary or Delete which defines it's colour.

Update all references to secondary button to use this new format

Update Delete order Page

Add / Update Delete order page tests

+++ src/NHSD.GPIT.BuyingCatalogue.WebApp/Areas/Order/Controllers/DeleteOrderController.cs
@@ -40,6 +40,17 @@ namespace NHSD.GPIT.BuyingCatalogue.WebApp.Areas.Order.Controllers
         [HttpPost]
         public async Task<IActionResult> DeleteOrder(string internalOrgId, CallOffId callOffId, DeleteOrderModel model)
         {
+            if (!ModelState.IsValid)
+                return View(model);
+
+            if (!model.SelectedOption.Value)
+            {
+                return RedirectToAction(
+                    nameof(OrderController.Order),
+                    typeof(OrderController).ControllerName(),
+                    new { internalOrgId, callOffId });
+            }
+
             await orderService.DeleteOrder(callOffId, internalOrgId);
 
             return RedirectToAction(
